move confirmation tests
