+++ e2e/specs/confirmations/transactions/approve-default-erc20.spec.js
@@ -1,30 +1,34 @@
 'use strict';
-import { SmokeConfirmations } from '../../tags';
-import TestHelpers from '../../helpers';
-import { loginToApp } from '../../viewHelper';
-import FixtureBuilder from '../../fixtures/fixture-builder';
+import { SmokeConfirmations } from '../../../tags';
+import TestHelpers from '../../../helpers';
+import { loginToApp } from '../../../viewHelper';
+import FixtureBuilder from '../../../fixtures/fixture-builder';
 import {
   withFixtures,
   defaultGanacheOptions,
-} from '../../fixtures/fixture-helper';
+} from '../../../fixtures/fixture-helper';
 
-import TabBarComponent from '../../pages/TabBarComponent';
-import TestDApp from '../../pages/Browser/TestDApp';
-import { SMART_CONTRACTS } from '../../../app/util/test/smart-contracts';
-import ContractApprovalBottomSheet from '../../pages/Browser/ContractApprovalBottomSheet';
-import Assertions from '../../utils/Assertions';
-import { ActivitiesViewSelectorsText } from '../../selectors/Transactions/ActivitiesView.selectors';
+import { SMART_CONTRACTS } from '../../../../app/util/test/smart-contracts';
+import { ContractApprovalBottomSheetSelectorsText } from '../../../selectors/Browser/ContractApprovalBottomSheet.selectors';
+import { ActivitiesViewSelectorsText } from '../../../selectors/Transactions/ActivitiesView.selectors';
+
+import ContractApprovalBottomSheet from '../../../pages/Browser/ContractApprovalBottomSheet';
+import Assertions from '../../../utils/Assertions';
+import TabBarComponent from '../../../pages/TabBarComponent';
+import TestDApp from '../../../pages/Browser/TestDApp';
 
 const HST_CONTRACT = SMART_CONTRACTS.HST;
+const EXPECTED_TOKEN_AMOUNT = '7';
 
-describe(SmokeConfirmations('ERC20 - Increase Allowance'), () => {
+describe(SmokeConfirmations('ERC20 tokens'), () => {
   beforeAll(async () => {
+    jest.setTimeout(170000);
     if (device.getPlatform() === 'android') {
       await TestHelpers.reverseServerPort();
     }
   });
 
-  it('from a dApp', async () => {
+  it('approve default ERC20 token amount from a dapp', async () => {
     await withFixtures(
       {
         dapp: true,
@@ -43,37 +47,36 @@ describe(SmokeConfirmations('ERC20 - Increase Allowance'), () => {
         await loginToApp();
         // Navigate to the browser screen
         await TabBarComponent.tapBrowser();
-
         await TestDApp.navigateToTestDappWithContract({
           contractAddress: hstAddress,
         });
-        await TestDApp.tapIncreaseAllowanceButton();
+        await TestDApp.tapApproveButton();
 
-        //Input custom token amount
         await Assertions.checkIfVisible(
           ContractApprovalBottomSheet.approveTokenAmount,
         );
-        await ContractApprovalBottomSheet.clearInput();
-        await ContractApprovalBottomSheet.inputCustomAmount('2');
 
-        // Assert that custom token amount is shown
         await Assertions.checkIfElementToHaveText(
           ContractApprovalBottomSheet.approveTokenAmount,
-          '2',
+          EXPECTED_TOKEN_AMOUNT,
         );
         // Tap next button
+        await Assertions.checkIfTextIsDisplayed(
+          ContractApprovalBottomSheetSelectorsText.NEXT,
+        );
         await ContractApprovalBottomSheet.tapNextButton();
 
+        await Assertions.checkIfTextIsDisplayed(
+          ContractApprovalBottomSheetSelectorsText.APPROVE,
+        );
         // Tap approve button
         await ContractApprovalBottomSheet.tapApproveButton();
 
         // Navigate to the activity screen
         await TabBarComponent.tapActivity();
 
-        // Assert that the ERC20 activity is an increase allowance and it is confirmed
-        await Assertions.checkIfTextIsDisplayed(
-          ActivitiesViewSelectorsText.INCREASE_ALLOWANCE_METHOD,
-        );
+        // Assert erc20 is approved
+
         await Assertions.checkIfTextIsDisplayed(
           ActivitiesViewSelectorsText.CONFIRM_TEXT,
         );
