+++ e2e/specs/confirmations/transactions/increase-allowance-erc20.spec.js
@@ -1,34 +1,30 @@
 'use strict';
-import { SmokeConfirmations } from '../../tags';
-import TestHelpers from '../../helpers';
-import { loginToApp } from '../../viewHelper';
-import FixtureBuilder from '../../fixtures/fixture-builder';
+import { SmokeConfirmations } from '../../../tags';
+import TestHelpers from '../../../helpers';
+import { loginToApp } from '../../../viewHelper';
+import FixtureBuilder from '../../../fixtures/fixture-builder';
 import {
   withFixtures,
   defaultGanacheOptions,
-} from '../../fixtures/fixture-helper';
+} from '../../../fixtures/fixture-helper';
 
-import { SMART_CONTRACTS } from '../../../app/util/test/smart-contracts';
-import { ContractApprovalBottomSheetSelectorsText } from '../../selectors/Browser/ContractApprovalBottomSheet.selectors';
-import { ActivitiesViewSelectorsText } from '../../selectors/Transactions/ActivitiesView.selectors';
-
-import ContractApprovalBottomSheet from '../../pages/Browser/ContractApprovalBottomSheet';
-import Assertions from '../../utils/Assertions';
-import TabBarComponent from '../../pages/TabBarComponent';
-import TestDApp from '../../pages/Browser/TestDApp';
+import TabBarComponent from '../../../pages/TabBarComponent';
+import TestDApp from '../../../pages/Browser/TestDApp';
+import { SMART_CONTRACTS } from '../../../../app/util/test/smart-contracts';
+import ContractApprovalBottomSheet from '../../../pages/Browser/ContractApprovalBottomSheet';
+import Assertions from '../../../utils/Assertions';
+import { ActivitiesViewSelectorsText } from '../../../selectors/Transactions/ActivitiesView.selectors';
 
 const HST_CONTRACT = SMART_CONTRACTS.HST;
-const EXPECTED_TOKEN_AMOUNT = '7';
 
-describe(SmokeConfirmations('ERC20 tokens'), () => {
+describe(SmokeConfirmations('ERC20 - Increase Allowance'), () => {
   beforeAll(async () => {
-    jest.setTimeout(170000);
     if (device.getPlatform() === 'android') {
       await TestHelpers.reverseServerPort();
     }
   });
 
-  it('approve default ERC20 token amount from a dapp', async () => {
+  it('from a dApp', async () => {
     await withFixtures(
       {
         dapp: true,
@@ -47,36 +43,37 @@ describe(SmokeConfirmations('ERC20 tokens'), () => {
         await loginToApp();
         // Navigate to the browser screen
         await TabBarComponent.tapBrowser();
+
         await TestDApp.navigateToTestDappWithContract({
           contractAddress: hstAddress,
         });
-        await TestDApp.tapApproveButton();
+        await TestDApp.tapIncreaseAllowanceButton();
 
+        //Input custom token amount
         await Assertions.checkIfVisible(
           ContractApprovalBottomSheet.approveTokenAmount,
         );
+        await ContractApprovalBottomSheet.clearInput();
+        await ContractApprovalBottomSheet.inputCustomAmount('2');
 
+        // Assert that custom token amount is shown
         await Assertions.checkIfElementToHaveText(
           ContractApprovalBottomSheet.approveTokenAmount,
-          EXPECTED_TOKEN_AMOUNT,
+          '2',
         );
         // Tap next button
-        await Assertions.checkIfTextIsDisplayed(
-          ContractApprovalBottomSheetSelectorsText.NEXT,
-        );
         await ContractApprovalBottomSheet.tapNextButton();
 
-        await Assertions.checkIfTextIsDisplayed(
-          ContractApprovalBottomSheetSelectorsText.APPROVE,
-        );
         // Tap approve button
         await ContractApprovalBottomSheet.tapApproveButton();
 
         // Navigate to the activity screen
         await TabBarComponent.tapActivity();
 
-        // Assert erc20 is approved
-
+        // Assert that the ERC20 activity is an increase allowance and it is confirmed
+        await Assertions.checkIfTextIsDisplayed(
+          ActivitiesViewSelectorsText.INCREASE_ALLOWANCE_METHOD,
+        );
         await Assertions.checkIfTextIsDisplayed(
           ActivitiesViewSelectorsText.CONFIRM_TEXT,
         );
