+++ e2e/specs/confirmations/transactions/set-approve-for-all-erc721.spec.js
@@ -1,21 +1,21 @@
 'use strict';
 
-import { SmokeConfirmations } from '../../tags';
-import TestHelpers from '../../helpers';
-import { loginToApp } from '../../viewHelper';
+import { SmokeConfirmations } from '../../../tags';
+import TestHelpers from '../../../helpers';
+import { loginToApp } from '../../../viewHelper';
 
-import TabBarComponent from '../../pages/TabBarComponent';
-import TestDApp from '../../pages/Browser/TestDApp';
-import FixtureBuilder from '../../fixtures/fixture-builder';
+import TabBarComponent from '../../../pages/TabBarComponent';
+import TestDApp from '../../../pages/Browser/TestDApp';
+import FixtureBuilder from '../../../fixtures/fixture-builder';
 import {
   withFixtures,
   defaultGanacheOptions,
-} from '../../fixtures/fixture-helper';
-import { SMART_CONTRACTS } from '../../../app/util/test/smart-contracts';
-import { ActivitiesViewSelectorsText } from '../../selectors/Transactions/ActivitiesView.selectors';
-import Assertions from '../../utils/Assertions';
-import { ContractApprovalBottomSheetSelectorsText } from '../../selectors/Browser/ContractApprovalBottomSheet.selectors';
-import ContractApprovalBottomSheet from '../../pages/Browser/ContractApprovalBottomSheet';
+} from '../../../fixtures/fixture-helper';
+import { SMART_CONTRACTS } from '../../../../app/util/test/smart-contracts';
+import { ActivitiesViewSelectorsText } from '../../../selectors/Transactions/ActivitiesView.selectors';
+import Assertions from '../../../utils/Assertions';
+import { ContractApprovalBottomSheetSelectorsText } from '../../../selectors/Browser/ContractApprovalBottomSheet.selectors';
+import ContractApprovalBottomSheet from '../../../pages/Browser/ContractApprovalBottomSheet';
 
 describe(SmokeConfirmations('ERC721 token'), () => {
   const NFT_CONTRACT = SMART_CONTRACTS.NFTS;
